% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World!}
\usage{
hello(text = world())
}
\arguments{
\item{text}{text string to be}
}
\value{
a character string with a '!' added to the end
}
\description{
More in-depth description goes here.
}
\examples{
hello()
# "world!"
hello("example")
# "example!"
}

